// app/src/renderer/api/gen/truss/v1/health_connect.ts
// @generated by protoc-gen-connect-es v1.6.1 with parameter "target=ts"
// @generated from file truss/v1/health.proto (package truss.v1, syntax proto3)
/* eslint-disable */
// @ts-nocheck

import { PingRequest, PingResponse } from "./health_pb.js";
import { MethodKind } from "@bufbuild/protobuf";

/**
 * @generated from service truss.v1.HealthService
 */
export const HealthService = {
  typeName: "truss.v1.HealthService",
  methods: {
    /**
     * @generated from rpc truss.v1.HealthService.Ping
     */
    ping: {
      name: "Ping",
      I: PingRequest,
      O: PingResponse,
      kind: MethodKind.Unary,
    },
  }
} as const;