// app/src/renderer/api/gen/truss/v1/helm_connect.ts
// @generated by protoc-gen-connect-es v1.6.1 with parameter "target=ts"
// @generated from file truss/v1/helm.proto (package truss.v1, syntax proto3)
/* eslint-disable */
// @ts-nocheck

import { GetReleaseHistoryRequest, GetReleaseHistoryResponse, GetReleaseRequest, GetReleaseResponse, GetReleaseValuesRequest, GetReleaseValuesResponse, ListReleasesRequest, ListReleasesResponse, RollbackReleaseRequest, RollbackReleaseResponse, UninstallReleaseRequest, UninstallReleaseResponse, UpgradeReleaseRequest, UpgradeReleaseResponse } from "./helm_pb.js";
import { MethodKind } from "@bufbuild/protobuf";

/**
 * @generated from service truss.v1.HelmService
 */
export const HelmService = {
  typeName: "truss.v1.HelmService",
  methods: {
    /**
     * @generated from rpc truss.v1.HelmService.ListReleases
     */
    listReleases: {
      name: "ListReleases",
      I: ListReleasesRequest,
      O: ListReleasesResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc truss.v1.HelmService.GetRelease
     */
    getRelease: {
      name: "GetRelease",
      I: GetReleaseRequest,
      O: GetReleaseResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc truss.v1.HelmService.GetReleaseValues
     */
    getReleaseValues: {
      name: "GetReleaseValues",
      I: GetReleaseValuesRequest,
      O: GetReleaseValuesResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc truss.v1.HelmService.GetReleaseHistory
     */
    getReleaseHistory: {
      name: "GetReleaseHistory",
      I: GetReleaseHistoryRequest,
      O: GetReleaseHistoryResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc truss.v1.HelmService.UninstallRelease
     */
    uninstallRelease: {
      name: "UninstallRelease",
      I: UninstallReleaseRequest,
      O: UninstallReleaseResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc truss.v1.HelmService.RollbackRelease
     */
    rollbackRelease: {
      name: "RollbackRelease",
      I: RollbackReleaseRequest,
      O: RollbackReleaseResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc truss.v1.HelmService.UpgradeRelease
     */
    upgradeRelease: {
      name: "UpgradeRelease",
      I: UpgradeReleaseRequest,
      O: UpgradeReleaseResponse,
      kind: MethodKind.Unary,
    },
  }
} as const;