// app/src/renderer/api/gen/truss/v1/yaml_connect.ts
// @generated by protoc-gen-connect-es v1.6.1 with parameter "target=ts"
// @generated from file truss/v1/yaml.proto (package truss.v1, syntax proto3)
/* eslint-disable */
// @ts-nocheck

import { ApplyYamlRequest, ApplyYamlResponse, DiffYamlRequest, DiffYamlResponse, GetYamlRequest, GetYamlResponse } from "./yaml_pb.js";
import { MethodKind } from "@bufbuild/protobuf";

/**
 * @generated from service truss.v1.YamlService
 */
export const YamlService = {
  typeName: "truss.v1.YamlService",
  methods: {
    /**
     * @generated from rpc truss.v1.YamlService.GetYaml
     */
    getYaml: {
      name: "GetYaml",
      I: GetYamlRequest,
      O: GetYamlResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc truss.v1.YamlService.ApplyYaml
     */
    applyYaml: {
      name: "ApplyYaml",
      I: ApplyYamlRequest,
      O: ApplyYamlResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc truss.v1.YamlService.DiffYaml
     */
    diffYaml: {
      name: "DiffYaml",
      I: DiffYamlRequest,
      O: DiffYamlResponse,
      kind: MethodKind.Unary,
    },
  }
} as const;